Remove template and move it all to index
